Split out tests into separate files, add watcher tests and simulate degradations in leases
